[Add] The generated fw is now up-to-date with the latest Xiaomi release
Now the init_sysroot.sh script looks for the file "home_vCAM_IDm" in each stock_firmware subfolder.
If it founds it, decryption occurs and the generated sysroots will be up to date with the above mentioned update file.

The Xiaomi's public key is not provided because of legal reasons. See the "/stock_firmware/common/README.md" file for further info on how to obtain it.